requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decouple==0.0.7
Django==4.1.6
django-filter==23.1
djangorestframework==3.14.0
fonttools==4.39.3
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
multimethod==1.9.1
networkx==3.1
numpy==1.23.5
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle @ file:///C:/b/abs_d77nxvklcq/croot/pycodestyle_1674267231034/work
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scipy==1.9.3
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
visions==0.7.5
wincertstore==0.2
ydata-profiling==4.1.2
